Module rust_jvm::model::class_file::access_flags::class_access_flags

Values of access flags for a class or interface.

Constants

ACC_ABSTRACT
Declared abstract; must not be instantiated.

ACC_ANNOTATION
Declared as an annotation type.

ACC_ENUM
Declared as an enum type.

ACC_FINAL
Declared final; no subclasses allowed.

ACC_INTERFACE
Is an interface, not a class.

ACC_PUBLIC
Declared public; may be accessed from outside its package.

ACC_SUPER
Treat superclass methods specially when invoked by the invokespecial instruction.

ACC_SYNTHETIC
Declared synthetic; not present in the source code.

Type Definitions

access_flag

t