Module rust_jvm::model::class_file::access_flags::parameter_access_flags

Values of access flags for parameters.

Constants

ACC_FINAL
Indicates that the formal parameter was declared final.

ACC_MANDATED
Indicates that the formal parameter was implicitly declared in source code, according to the specification of the language in which the source code was written (JLS 13.1). (The formal parameter is mandated by a language specification, so all compilers for the language must emit it.)

ACC_SYNTHETIC
Indicates that the formal parameter was not explicitly or implicitly declared in source code, according to the specification of the language in which the source code was written (JLS 13.1). (The formal parameter is an implementation artifact of the compiler which produced this class file.)

Type Definitions

access_flag

t