Module rust_jvm::model::class_file::attribute::stack_map_frame

Reexports

pub use self::verification_type_info::VerificationTypeInfo;

Modules

verification_type_info

Enums

StackMapFrame
A StackMapFrame variant stores a relative bytecode offset, the verification types (4.10.1.2) for the local variables, and the verification types for the operand stack. Each variant stores a bytecode offset relative to the previous StackMapFrame. The actual bytecode offset can be calculated as described in (4.7.4).

Tag