Enum rust_jvm::model::class_file::attribute::stack_map_frame::verification_type_info::VerificationTypeInfo

pub enum VerificationTypeInfo {
    Top,
    Integer,
    Float,
    Long,
    Double,
    Null,
    UninitializedThis,
    Object {
        class_index: constant_pool_index,
    },
    Uninitialized {
        offset: u2,
    },
}

Variants

Top

Integer

Float

Long

Double

Null

UninitializedThis

Object
Fields

class_index

Uninitialized
Fields

offset

The offset in the code array of the Code attribute that contains this StackMapTable attribute, of the new instruction that created the object stored in the location.