Module rust_jvm::parser

Contains a parser for a Java class file.

Examples

Basic usage: let data = include_bytes!("../../data/HelloWorld.class"); assert!(parse_class_file(data).is_done()); // returns a nom::IResult

Modules

class_file
A parser for a Java class file.

nom_support
Workaround to get sane nested errors from nom. The nom library does not provide parsers that can have controllable backtracking behavior, so we need to hack it in. Follows #160.