Module rust_jvm::parser::class_file

A parser for a Java class file.

Enums

Error

Functions

annotation

annotations

attribute

attribute_info

attribute_info_switch

class_file_parser

cp_index

cp_index_tag
Parses for a constant pool index and verifies that its the entry in the constant pool matches the specified tag.

cp_info

cp_info_info

cp_info_tag

element_value

element_value_pair

element_value_pairs

exception_table

field

inner_class

line_number_info

local_variable_info

local_variable_type_info

magic

maybe_cp_index_tag
Parses for a constant pool index that might be zero and verifies that its the entry in the constant pool matches the specified tag.

method

method_parameter

modified_utf8

parameter_annotations

parse_class_file
Parses a Java class file.

reference

reference_kind

stack_map_frame

stack_map_frame_info

target_info

type_annotation

type_annotations

type_path

verification_type_info

verification_type_info_tag

Type Definitions

ConstantPoolIndex
The type of an index into the class file constant pool.

Input
The input type to the parser.

ParseResult
The return type of a backtracking-controllable parser method.