Module rust_jvm::parser::nom_support

Workaround to get sane nested errors from nom. The nom library does not provide parsers that can have controllable backtracking behavior, so we need to hack it in. Follows #160.

This module contains support for converting to and from nom parsers and backtracking-controllable parsers. A parser that can be controlled for backtracking has type std::result::Result<nom::IResult, nom::Err<I, E>>, while a nom parser has type nom::IResult.