Module rust_jvm::vm

The public interface for the Java virtual machine.

Modules

bytecode
The Java bytecode.

class
Internal JVM representations of classes and methods.

class_loader

constant_pool
The runtime constant pool.

frame
The runtime state of a method that is currently executing in the virtual machine.

native

sig
Signatures of runtime constant pool entities that serve to uniquely identify those entities. These are derived from structures in the binary representation of the constant pool (5.1).

symref
A symbolic reference to an entity in the runtime constant pool (5.1). Symbolic references must be resolved (5.4.3) before their usage by the interpreter.

value
Internal JVM representations of Java values.

Structs

VirtualMachine
The top-level virtual machine. The virtual machine contains a reference to its bootstrap class loader, which is used to load the main class and all of that class's dependencies.