Enum rust_jvm::vm::class::MethodCode

enum MethodCode {
    Concrete {
        max_locals: u16,
        code: Vec<u8>,
        exception_table: Vec<ExceptionTableEntry>,
    },
    Abstract,
    Native(NativeMethod),
    NativeNotFound,
}

A representation of the code associated with a method, or more generally, the action that should be taken when a method is invoked.

Variants

Concrete

The code for a non-abstract, non-native Java method. Such contains executable bytecode which may be used to create a new JVM stack frame.

Fields

max_locals

code

exception_table

Abstract

to invoke an abstract method fails with AbstractMethodError.

Native

The code for a native Java method for which the class loader has located a corresponding Rust function pointer.

NativeNotFound

The code for a native Java method for which the class loader failed to locate a Rust function pointer.