Struct rust_jvm::vm::class::Method

pub struct Method {
    pub symref: Method,
    pub access_flags: u16,
    code: MethodCode,
}

A JVM representation of a method in a loaded class.

Fields

symref

The method's signature, comprised of its name and argument and return types.

access_flags

The method's access flags.

code

A MethodCode variant, which is used to actually invoke the method.

Methods

impl Method

fn new(symref: Method, method_info: MethodInfo) -> Self

fn invoke(&self, class: &Class, class_loader: &mut ClassLoader, args: Vec<Value>) -> Option<Value>