Enum rust_jvm::vm::class_loader::Error

pub enum Error {
    ClassNotFound {
        name: String,
        error: Error,
    },
    ClassFormat,
    UnsupportedVersion {
        major: u16,
        minor: u16,
    },
    NoClassDefFound {
        name: String,
    },
    NoClassDefFoundCause {
        name: String,
        not_found: String,
    },
    IncompatibleClassChange(String),
    ClassCircularity,
}

Variants

ClassNotFound

If no "purported representation" of the class is found. 5.3.1.

Fields

name

error

ClassFormat

The "purported representation" does not follow the class file format. 5.3.5.

UnsupportedVersion

The "purported representation" is not of a supported version. 5.3.5.

Fields

major

minor

NoClassDefFound

The "purported representation" does not actually represent the requested class. 5.3.5.

Fields

name

NoClassDefFoundCause

(A subtlety here is that recursive class loading to load superclasses is performed as part of resolution (5.3.5, step 3). Therefore, a ClassNotFoundException that results from a class loader failing to load a superclass must be wrapped in a NoClassDefFoundError.) 5.3

Fields

name

not_found

IncompatibleClassChange

ClassCircularity

The class is its own superclass or superinterface. 5.3.5.