Struct rust_jvm::vm::constant_pool::RuntimeConstantPool

pub struct RuntimeConstantPool {
    entries: OneIndexedVec<Option<RuntimeConstantPoolEntry>>,
}

A runtime constant pool. This just consists of a OneIndexedVec of constant pool entries.

Fields

entries

Methods

impl RuntimeConstantPool

fn new(constant_pool: &ConstantPool) -> Self

Creates a new runtime constant pool from the ConstantPool returned by the class file parser. Most of this process involves constructing sig and symref structures representing the symbolic references in the constant pool.

fn force_class_ref(constant_pool: &ConstantPool, info: &ConstantPoolInfo) -> Class

Constructs a symref::Class from a ConstantPoolInfo::Class, panicking if info is of a different variant of ConstantPoolInfo.

This should only be called where the specification requires that info be of the correct variant.

fn force_name_and_type(constant_pool: &ConstantPool, info: &ConstantPoolInfo) -> (String, String)

Constructs a tuple of name and descriptor (type) strings from a ConstantPoolInfo::NameAndType, panicking if info is of a different variant of ConstantPoolInfo. The names of classes are binary names (4.2.1) while the names of fields and methods are unqualified names (4.2.2). Descriptor formats vary depending on the type of descriptor being referenced (4.3).

This should only be called where the specification requires that info be of the correct variant.

fn force_string(info: &ConstantPoolInfo) -> ModifiedUtf8String

Constructs a ModifiedUtf8String from a ConstantPoolInfo::Utf8, panicking in info is of a different variant of ConstantPoolInfo.

This should only be called where the specification requires that info be of the correct variant.

fn lookup_raw_string(&self, index: constant_pool_index) -> String

Returns the String at the runtime constant pool entry at index, panicking if that entry is not a RuntimeConstantPoolEntry::StringValue. This is used during class creation, because the structures describing fields and methods later in the class file (after the constant pool) use constant pool indices to refer to their names.

fn resolve_literal(&self, index: constant_pool_index, class_loader: &mut ClassLoader) -> Result<Value, Error>

Resolves a literal value in the constant pool into a Value. For String literals, this requires instantiating an instance of the String class, which we do by calling the String(char[]) constructor using the content of the modified UTF-8 string in the constant pool, parsed into UTF-16.