Struct rust_jvm::vm::VirtualMachine

pub struct VirtualMachine {
    bootstrap_class_loader: ClassLoader,
}

The top-level virtual machine. The virtual machine contains a reference to its bootstrap class loader, which is used to load the main class and all of that class's dependencies.

The virtual machine's start method causes initialization of the boostrap class loader, which loads the main class and invokes the main(String[]) method to begin execution of the program. The virtual machine terminates either when the program completes successfully, or when there is an uncaught error in the program or the internal implementation of the virtual machine itself.

Fields

bootstrap_class_loader

The bootstrap class loader used to initialize the virtual machine.

Methods

impl VirtualMachine

fn new() -> Self

fn start(self, main_class: Class)

Begin execution of the virtual machine instance's main(String[]) method.