Module rust_jvm::vm::symref

A symbolic reference to an entity in the runtime constant pool (5.1). Symbolic references must be resolved (5.4.3) before their usage by the interpreter.

Structs

Class
A symbolic reference to a class constant pool entry.

Field
A symbolic reference to a field constant pool entry.

Method
A symbolic reference to a method constant pool entry.