Enum rust_jvm::vm::value::Value

pub enum Value {
    Int(Wrapping<i32>),
    Float(f32),
    Long(Wrapping<i64>),
    Double(f64),
    ScalarReference(Rc<RefCell<Scalar>>),
    ArrayReference(Rc<RefCell<Array>>),
    NullReference,
}

A value in the Java virtual machine.

Variants

Int

A 32-bit signed integral type, representing the Java types byte, char, short, int, and boolean.

Float

A 32-bit floating-point type, representing the Java type float.

Long

A 64-bit signed integral type, representing the Java type long.

Double

A 64-bit floating-point type, representing the Java type double.

ScalarReference

A reference to a scalar Java object in the heap.

ArrayReference

A reference to a Java array in the heap.

NullReference

A reference to a Java object which is null.